script/toolbox/requirements.txt
python-dotenv==1.2.1
psycopg2-binary==2.9.11
requests==2.32.5
